backend/pyproject.toml
[project]
name = "inspector-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "bcrypt==4.0.1",
    "python-multipart",
    "httpx",
    "aioboto3",
    "arq",
    "Pillow",
    "WeasyPrint",
    "openpyxl",
    "pydantic-settings",
    "email-validator",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "mypy",
    "ruff",
    "aiosqlite",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
